backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
supabase>=2.11.0
anthropic>=0.42.0
httpx>=0.28.0
PyJWT[crypto]>=2.10.0
google-api-python-client>=2.160.0
google-auth-oauthlib>=1.2.0
python-multipart>=0.0.18
pytest>=8.3.0
pytest-asyncio>=0.25.0
